Break apart invoice structs
