+++ src/lib.rs
@@ -7,11 +7,15 @@ pub use dim::ucum;
 mod billable_item;
 mod board_dimensions;
 mod invoice;
+mod invoice_summary;
 mod lumber_types;
+mod order_info;
 mod site_info;
 
 pub use self::billable_item::BillableItem;
 pub use self::board_dimensions::BoardDimensions;
-pub use self::invoice::{Invoice, OrderInfo, Summary};
+pub use self::invoice::Invoice;
+pub use self::invoice_summary::InvoiceSummary;
 pub use self::lumber_types::LumberType;
+pub use self::order_info::OrderInfo;
 pub use self::site_info::SiteInfo;
